feat(auth): implement auth controller, routes and service
- Added controller methods for register, login, logout, refresh
- Set up auth routes using Express Router
- AuthService handles business logic including JWT token generation
