+++ server/src/middleware/errorHandler.js
@@ -1,0 +1,8 @@
+const errorHandler = (err, req, res, next) => {
+  res.status(err.status || 500).json({
+    success: false,
+    message: err.message || "Internal Server Error",
+  });
+};
+
+module.exports = errorHandler;
